refactor(app/core): outline `ResponseBody` middleware
we'll add a few tests for this middleware shortly.

this commit moves this middleware out into its own submodule.
